+++ plugins/dialog/android/src/main/java/DialogPlugin.kt
@@ -5,7 +5,6 @@
 package app.tauri.dialog
 
 import android.app.Activity
-import android.app.AlertDialog
 import android.content.Intent
 import android.net.Uri
 import android.os.Handler
@@ -21,6 +20,7 @@ import app.tauri.plugin.Invoke
 import app.tauri.plugin.JSArray
 import app.tauri.plugin.JSObject
 import app.tauri.plugin.Plugin
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 
 @InvokeArg
 class Filter {
@@ -158,7 +158,7 @@ class DialogPlugin(private val activity: Activity): Plugin(activity) {
 
     Handler(Looper.getMainLooper())
       .post {
-        val builder = AlertDialog.Builder(activity)
+        val builder = MaterialAlertDialogBuilder(activity)
 
         if (args.title != null) {
           builder.setTitle(args.title)
@@ -191,8 +191,7 @@ class DialogPlugin(private val activity: Activity): Plugin(activity) {
           }
         }
 
-        val dialog = builder.create()
-        dialog.show()
+        builder.show()
       }
   }
 
@@ -205,9 +204,10 @@ class DialogPlugin(private val activity: Activity): Plugin(activity) {
       val intent = Intent(Intent.ACTION_CREATE_DOCUMENT)
       intent.addCategory(Intent.CATEGORY_OPENABLE)
       intent.putExtra(Intent.EXTRA_TITLE, args.fileName ?: "")
-      intent.type = "*/*"
-
-      if (parsedTypes.isNotEmpty()) {
+      if (parsedTypes.size == 1) {
+        intent.type = parsedTypes.first()
+      } else {
+        intent.type = "*/*"
         intent.putExtra(Intent.EXTRA_MIME_TYPES, parsedTypes)
       }