+++ plugins/notification/android/src/main/java/NotificationStorage.kt
@@ -80,11 +80,11 @@ class NotificationStorage(private val context: Context, private val jsonMapper: 
 
   fun writeActionGroup(actions: List<ActionType>) {
     for (type in actions) {
-      val i = type.id
       val editor = getStorage(ACTION_TYPES_ID + type.id).edit()
       editor.clear()
       editor.putInt("count", type.actions.size)
-      for (action in type.actions) {
+      for (i in 0 until type.actions.size) {
+        val action = type.actions[i]
         editor.putString("id$i", action.id)
         editor.putString("title$i", action.title)
         editor.putBoolean("input$i", action.input ?: false)