+++ plugins/deep-link/examples/app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.2.13]
+
+### Dependencies
+
+- Upgraded to `deep-link-js@2.4.10`
+
 ## \[2.2.12]
 
 ### Dependencies