+++ plugins/websocket/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.4.3]
+
+- [`f8053e65`](https://github.com/tauri-apps/plugins-workspace/commit/f8053e659e4ccd85c1f52833411ff8417cbc5e69) ([#3527](https://github.com/tauri-apps/plugins-workspace/pull/3527) by [@Legend-Master](https://github.com/tauri-apps/plugins-workspace/../../Legend-Master)) Documented Cargo feature flags in each plugin's crate-level documentation.
+
 ## \[2.4.2]
 
 ### bug